Backend/genai exchange chatbot backend/requirements.txt
# Web server
fastapi==0.116.1
uvicorn==0.18.3
python-multipart==0.0.9

# Google Cloud clients
google-cloud-speech==2.26.0
google-cloud-translate==3.16.0
google-cloud-texttospeech==2.16.4
google-cloud-aiplatform==1.71.1   # For Vertex AI + RAG

# Generative AI (Gemini API)
google-generativeai==0.8.5

# Core dependencies
protobuf==4.25.5
grpcio==1.74.0
requests==2.32.3

# Runtime libraries
pydantic==2.11.7
numpy==1.26.4
httpcore==1.0.9
h11==0.16.0


